Move Google Autenticator to lib
Move the Google Authenticator validation code from every db module to
lib/TOTPValidate instead.
Also minor helper cleanup.
